[refactor/#34] filiter 패키지 구조 변경
