+++ app/src/main/java/com/smashing/app/presentation/search/component/MatchingSearchFilterChip.kt
@@ -11,28 +11,34 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.smashing.app.R.drawable.ic_arrow_down
+import com.smashing.app.R.drawable.ic_close_sm
 import com.smashing.app.core.designsystem.theme.SmashingAndroidTheme
+import com.smashing.app.core.designsystem.theme.SmashingTheme
 import com.smashing.app.core.designsystem.theme.SmashingTheme.typography
 import com.smashing.app.core.designsystem.theme.gray300
 import com.smashing.app.core.extension.noRippleClickable
-import com.smashing.app.presentation.search.style.FilterType
-
+import com.smashing.app.presentation.search.style.FilterStyle
+import com.smashing.app.presentation.search.style.FilterStyle.DEFAULT
+import com.smashing.app.presentation.search.style.FilterStyle.VARIANT
+import com.smashing.app.presentation.search.style.FilterStyleData
 
 @Composable
 fun MatchingSearchFilterChip(
-    type: FilterType,
+    style: FilterStyle,
     text: String,
     onFilterClick: () -> Unit,
     onFilterDelete: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val filterStyle = type.getStyle()
+    val filterStyle = style.getStyle()
 
     Row(
         modifier = modifier
@@ -56,7 +62,7 @@ fun MatchingSearchFilterChip(
         Icon(
             imageVector = ImageVector.vectorResource(filterStyle.iconRes),
             contentDescription = null,
-            modifier = if(type == FilterType.VARIANT) {
+            modifier = if (style == VARIANT) {
                 Modifier.noRippleClickable(onClick = onFilterDelete)
             } else {
                 Modifier
@@ -64,6 +70,24 @@ fun MatchingSearchFilterChip(
             tint = filterStyle.iconTint,
         )
     }
+}
+
+@ReadOnlyComposable
+@Composable
+private fun FilterStyle.getStyle(): FilterStyleData = when (this) {
+    DEFAULT -> FilterStyleData(
+        bgColor = SmashingTheme.colors.bgSurfacePressed,
+        txtColor = SmashingTheme.colors.txtPrimary,
+        iconRes = ic_arrow_down,
+        iconTint = SmashingTheme.colors.iconPrimary,
+    )
+
+    VARIANT -> FilterStyleData(
+        bgColor = SmashingTheme.colors.bgSelected,
+        txtColor = SmashingTheme.colors.txtPrimaryReverse,
+        iconRes = ic_close_sm,
+        iconTint = SmashingTheme.colors.iconTertiary,
+    )
 }
 
 
@@ -77,7 +101,7 @@ private fun MatchingSearchFilterChipPreview() {
                 .padding(5.dp),
         ) {
             MatchingSearchFilterChip(
-                type = FilterType.DEFAULT,
+                style = DEFAULT,
                 text = "티어",
                 onFilterClick = {},
                 onFilterDelete = {},
@@ -86,7 +110,7 @@ private fun MatchingSearchFilterChipPreview() {
             Spacer(modifier = Modifier.height(10.dp))
 
             MatchingSearchFilterChip(
-                type = FilterType.VARIANT,
+                style = VARIANT,
                 text = "브론즈",
                 onFilterClick = {},
                 onFilterDelete = {},
